Update to fix negative item IDs for things like spring water actors.

+++ code/include/game/player.h
@@ -218,7 +218,7 @@ struct Player : public Actor {
   u8 gap_904[8];
   char field_914[4];
   u32 door;
-  u32 get_item_id; // Index+1 to the get item ID table.
+  s8 get_item_id; // Index+1 to the get item ID table.
   Actor* grabbable_actor;
   u8 gap_924[4];
   u32 field_928; // Epona Actor
